remove invalid profiling options from cabal file
`-auto-exported` is not a valid GHC flag, causing cabal builds with profiling enabled to fail.

Instead of changing it to `-fprof-auto-exported`, I have instead removed it from the cabal
file. Cabal's documentation, linked below, recommends that packages do not choose their
own profiling options, so that the person building the package can decide.

https://www.haskell.org/cabal/users-guide/developing-packages.html#pkg-field-ghc-prof-options